tests/wpt/web-platform-tests/referrer-policy/css-integration/image/internal-stylesheet.html

<!DOCTYPE html>
<html>
  <head>
    <title>CSS integration - image from internal stylesheet</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/common/utils.js"></script>
    <!-- Common global functions for referrer-policy tests. -->
    <script src="/referrer-policy/generic/common.js"></script>
    <meta name="referrer" content="origin">
  </head>
  <body>
    <p>Check that resources from internal stylesheets are loaded with
    the referrer and referrer policy from the document.</p>

    <div class="styled"></div>

    <script>
      async_test(function(css_test) {
        var id = token();
        var css_url = location.protocol + "//www1." + location.hostname + ":" + location.port + "/referrer-policy/generic/subresource/image.py" + "?id=" + id;
        var img_url = css_url + "&report-headers";

        var style = document.createElement("style");
        style.type = 'text/css';
        style.appendChild(document.createTextNode("div.styled::before { content:url(" + css_url + ")}"));
        document.head.appendChild(style);
        css_test.step_timeout(function() {
          queryXhr(img_url, function(message) {
            assert_own_property(message, "headers");
            assert_own_property(message, "referrer");
            assert_equals(message.referrer, location.origin + "/");
            css_test.done();
          }, null, null, css_test);
        }, 1000);
      }, "Image from internal stylesheet.");
    </script>

    <div id="log"></div>
  </body>
</html>
